pyproject.toml
[tool.poetry]
name = "study-with-ai"
version = "0.1.0"
description = ""
authors = ["yoo-sh-96 <tmdguq90051@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.116.1"}
pydantic-settings = "^2.10.1"
tortoise-orm = "^0.25.1"
aerich = "^0.9.1"
asyncpg = "^0.30.0"
python-dotenv = "^1.1.1"
orjson = "^3.11.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.17.1"
black = "^25.1.0"
ruff = "^0.12.7"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "app.configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
